backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
httpx
meilisearch
sentence-transformers
pillow
torch
openai
transformers
sentencepiece
protobuf
